Docker: Install package 'ca-certificates'
This package is necessary to be able to verify certificates when communicating with the Concordium Node over HTTPS.

Installing the package is actually redundant as it's already included in the currently used base image. But this may change in the future, and explicitly installing it communicates that we depend on it.