Introduced batches to file m0037
It was found that the running of this took a long time.
Introducing batches to the script and adding more logging
to show the progress.
The overall running of this particular script would still be
quite a while, expect possibly a few days. The log should print
how many rows are processed at the time of running.

Jira-Id: COR-1652